API Change!  Instead of two events "on_header_field" and "on_header_value", emit these as a single "on_header" event that is passed in two arguments.  This API should be slightly easier to use.
I also removed the redundant cb_id field from lhttp_parser so the
overhead of a parser was reduced by 4 bytes.  However, to facilitate
calling a 3 argument function, the procol between the lua
implementation of "execute" and the C implementation had to change.
The new protocol is to push a count of how many items are on the Lua
stack followed by the function followed by the arguments to that
function.

After this change the benchmarks are about the same even though a
function call was eliminated.